# internal/server/api/docs/openapi.yaml
openapi: 3.0.3
info:
  title: lanweave API
  version: 1.0.0
  description: |
    Control-plane REST API of the lanweave relay server.

    Authentication: most endpoints require a bearer access token obtained from
    `POST /api/v1/login` (renewable via `POST /api/v1/refresh`). Admin-only
    endpoints additionally require the account to be an administrator.

    Errors: every error response uses the single JSON envelope
    `{"error": "<machine_code>", "message": "<human sentence>"}`. The machine
    codes an operation can return are listed per response via `x-error-codes`.
    Two errors can occur on any operation and are not repeated in prose:
    `rate_limited` (429, global rate limiter) and `internal_error` (500).
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: /
...
paths:
...
  /api/v1/login:
    post:
      tags: [auth]
      operationId: login
      security: []
      summary: Log in with username and password
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/LoginRequest" }
      responses:
        "200":
          description: Credentials accepted; tokens issued.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/LoginResponse" }
        "400":
          description: Malformed body or missing fields.
          x-error-codes: [validation_error]
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
        "401":
          description: Unknown username or wrong password.
          x-error-codes: [invalid_credentials]
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
        "429": { $ref: "#/components/responses/RateLimited" }
        "500": { $ref: "#/components/responses/InternalError" }
  /api/v1/register:
    post:
      tags: [auth]
      operationId: register
      security: []
      summary: Create an account with an invite code
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/RegisterRequest" }
      responses:
        "201":
          description: Account created.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/RegisterResponse" }
        "400":
          description: Malformed body, bad username length, or password not meeting the complexity policy (8-64 ASCII characters with at least one uppercase letter, one lowercase letter and one digit).
          x-error-codes: [validation_error]
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
        "409":
          description: Username already taken.
          x-error-codes: [username_taken]
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
        "422":
          description: Invite code unknown, already used, or expired.
          x-error-codes: [invite_invalid]
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
        "429": { $ref: "#/components/responses/RateLimited" }
        "500": { $ref: "#/components/responses/InternalError" }
  /api/v1/refresh:
    post:
      tags: [auth]
      operationId: refresh
      security: []
      summary: Exchange a refresh token for a new access token
      description: Public on purpose — the access token is expired at refresh time; the refresh token in the body authenticates the call.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/RefreshRequest" }
      responses:
        "200":
          description: New access token issued; the refresh token keeps sliding.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/RefreshResponse" }
        "400":
          description: Malformed body or missing refresh token.
          x-error-codes: [validation_error]
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
        "401":
          description: Refresh token invalid, revoked or expired.
          x-error-codes: [invalid_refresh_token]
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
        "429": { $ref: "#/components/responses/RateLimited" }
        "500": { $ref: "#/components/responses/InternalError" }
  /api/v1/logout:
    post:
      tags: [auth]
      operationId: logout
      security: []
      summary: Revoke a refresh token
      description: Public for the same reason as refresh; revocation is idempotent.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/LogoutRequest" }
      responses:
        "204":
          description: Token revoked (or was already revoked).
        "400":
          description: Malformed body or missing refresh token.
          x-error-codes: [validation_error]
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
        "429": { $ref: "#/components/responses/RateLimited" }
        "500": { $ref: "#/components/responses/InternalError" }
  /api/v1/me:
    get:
      tags: [auth]
      operationId: getMe
      summary: Describe the authenticated account
      security: [{ bearerAuth: [] }]
      responses:
        "200":
          description: Current account.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/MeResponse" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "429": { $ref: "#/components/responses/RateLimited" }
        "500": { $ref: "#/components/responses/InternalError" }
  /api/v1/server:
    get:
      tags: [system]
      operationId: getServerInfo
      summary: WireGuard connection parameters
      description: Everything a client needs to configure its tunnel side.
      security: [{ bearerAuth: [] }]
      responses:
        "200":
          description: Server-side tunnel parameters.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ServerInfoResponse" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "429": { $ref: "#/components/responses/RateLimited" }
        "500": { $ref: "#/components/responses/InternalError" }
  /api/v1/nodes:
    post:
      tags: [nodes]
      operationId: registerNode
      summary: Register a device and allocate its VPN address
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/RegisterNodeRequest" }
      responses:
        "201":
          description: Device registered; the lowest free pool address was assigned and the WireGuard peer installed.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/NodeResponse" }
        "400":
          description: Malformed body, bad node name, or invalid WireGuard public key.
          x-error-codes: [validation_error]
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "409":
          description: Duplicate node name (per account), public key already registered, or the per-user device limit is reached.
          x-error-codes: [node_name_taken, pubkey_taken, device_limit_reached]
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
        "503":
          description: The address pool is exhausted.
          x-error-codes: [pool_exhausted]
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
        "429": { $ref: "#/components/responses/RateLimited" }
        "500": { $ref: "#/components/responses/InternalError" }
    get:
      tags: [nodes]
      operationId: listNodes
      summary: List the caller's devices
      security: [{ bearerAuth: [] }]
      responses:
        "200":
          description: All devices owned by the caller, with online status.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/NodeListResponse" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "429": { $ref: "#/components/responses/RateLimited" }
        "500": { $ref: "#/components/responses/InternalError" }
  /api/v1/nodes/{id}:
    delete:
      tags: [nodes]
      operationId: deleteNode
      summary: Delete one of the caller's devices
      description: Releases the VPN address, removes the WireGuard peer and drops all zone memberships of the device.
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: "#/components/parameters/NodeID"
      responses:
        "204":
          description: Device deleted.
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404":
          description: No such device, or it belongs to another account.
          x-error-codes: [not_found]
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
        "429": { $ref: "#/components/responses/RateLimited" }
        "500": { $ref: "#/components/responses/InternalError" }
  /api/v1/zones:
    post:
      tags: [zones]
      operationId: createZone
      summary: Create a zone (optionally auto-joining one of the caller's devices)
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreateZoneRequest" }
      responses:
        "201":
          description: Zone created; the caller is its owner.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ZoneResponse" }
        "400":
          description: Malformed body, bad zone name, or zone password shorter than 8 characters.
          x-error-codes: [validation_error]
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404":
          description: The auto-join node_id does not exist or belongs to another account (the zone is not created).
          x-error-codes: [not_found]
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
        "409":
          description: Zone name already taken, or the per-user owned-zone limit is reached.
          x-error-codes: [zone_name_taken, zone_limit_reached]
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
        "429": { $ref: "#/components/responses/RateLimited" }
        "500": { $ref: "#/components/responses/InternalError" }
    get:
      tags: [zones]
      operationId: listZones
      summary: List zones the caller participates in
      security: [{ bearerAuth: [] }]
      responses:
        "200":
          description: Zones the caller owns or has a device in.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ZoneListResponse" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "429": { $ref: "#/components/responses/RateLimited" }
        "500": { $ref: "#/components/responses/InternalError" }
  /api/v1/zones/{name}/join:
    post:
      tags: [zones]
      operationId: joinZone
      summary: Join a zone with one of the caller's devices
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: "#/components/parameters/ZoneName"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/JoinZoneRequest" }
      responses:
        "200":
          description: Device joined; traffic to other members is allowed immediately.
        "400":
          description: Malformed body or missing node_id/password.
          x-error-codes: [validation_error]
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403":
          description: Zone does not exist or the password is wrong (deliberately indistinguishable).
          x-error-codes: [invalid_zone_or_password]
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
        "404":
          description: The device does not exist or belongs to another account.
          x-error-codes: [not_found]
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
        "429": { $ref: "#/components/responses/RateLimited" }
        "500": { $ref: "#/components/responses/InternalError" }
  /api/v1/zones/{name}/leave:
    post:
      tags: [zones]
      operationId: leaveZone
      summary: Remove one of the caller's devices from a zone
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: "#/components/parameters/ZoneName"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/LeaveZoneRequest" }
      responses:
        "204":
          description: Device left the zone.
        "400":
          description: Malformed body or missing node_id.
          x-error-codes: [validation_error]
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404":
          description: Zone, device or membership not found (for the caller).
          x-error-codes: [not_found]
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
        "429": { $ref: "#/components/responses/RateLimited" }
        "500": { $ref: "#/components/responses/InternalError" }
  /api/v1/zones/{name}/members:
    get:
      tags: [zones]
      operationId: listZoneMembers
      summary: List a zone's members
      description: Full transparency within the zone — any member sees every member's device name, address and owner.
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: "#/components/parameters/ZoneName"
      responses:
        "200":
          description: Member devices of the zone.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ZoneMembersResponse" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404":
          description: Zone not found, or the caller is not a participant.
          x-error-codes: [not_found]
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
        "429": { $ref: "#/components/responses/RateLimited" }
        "500": { $ref: "#/components/responses/InternalError" }
  /api/v1/zones/{name}:
    patch:
      tags: [zones]
      operationId: changeZonePassword
      summary: Change a zone's password (owner only)
      description: Existing members are not evicted; only future joins need the new password.
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: "#/components/parameters/ZoneName"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ChangeZonePasswordRequest" }
      responses:
        "200":
          description: Password changed.
        "400":
          description: Malformed body or zone password shorter than 8 characters.
          x-error-codes: [validation_error]
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403":
          description: The caller is not the zone owner.
          x-error-codes: [forbidden]
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
        "404":
          description: Zone not found.
          x-error-codes: [not_found]
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
        "429": { $ref: "#/components/responses/RateLimited" }
        "500": { $ref: "#/components/responses/InternalError" }
    delete:
      tags: [zones]
      operationId: deleteZone
      summary: Delete a zone (owner only)
      description: Destroys the zone, all memberships and its firewall state; the name becomes available again.
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: "#/components/parameters/ZoneName"
      responses:
        "204":
          description: Zone deleted.
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403":
          description: The caller is not the zone owner.
          x-error-codes: [forbidden]
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
        "404":
          description: Zone not found.
          x-error-codes: [not_found]
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
        "429": { $ref: "#/components/responses/RateLimited" }
        "500": { $ref: "#/components/responses/InternalError" }
  /api/v1/zones/{name}/members/{node_id}:
    delete:
      tags: [zones]
      operationId: kickZoneMember
      summary: Kick a member device from a zone (owner only)
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: "#/components/parameters/ZoneName"
        - name: node_id
          in: path
          required: true
          description: ID of the member device to remove.
          schema: { type: integer, format: int64 }
      responses:
        "204":
          description: Member removed; its traffic to the zone stops immediately.
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403":
          description: The caller is not the zone owner.
          x-error-codes: [forbidden]
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
        "404":
          description: Zone, device or membership not found.
          x-error-codes: [not_found]
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
        "429": { $ref: "#/components/responses/RateLimited" }
        "500": { $ref: "#/components/responses/InternalError" }
  /api/v1/admin/invites:
    post:
      tags: [admin]
      operationId: createInvite
      summary: Mint an invite code (admin only)
      description: No request body. The code's expiry follows the server's invite_ttl configuration.
      security: [{ bearerAuth: [] }]
      responses:
        "201":
          description: Invite code created.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/CreateInviteResponse" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/ForbiddenAdmin" }
        "429": { $ref: "#/components/responses/RateLimited" }
        "500": { $ref: "#/components/responses/InternalError" }
    get:
      tags: [admin]
      operationId: listInvites
      summary: List invite codes (admin only)
      security: [{ bearerAuth: [] }]
      responses:
        "200":
          description: All invite codes with usage status.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/InviteListResponse" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/ForbiddenAdmin" }
        "429": { $ref: "#/components/responses/RateLimited" }
        "500": { $ref: "#/components/responses/InternalError" }
  /api/v1/admin/users/{id}:
    delete:
      tags: [admin]
      operationId: deleteUser
      summary: Delete a user and cascade all their state (admin only)
      description: Removes the user's devices (addresses freed, peers removed), their zone memberships, and every zone they own.
      security: [{ bearerAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          description: ID of the user to delete.
          schema: { type: integer, format: int64 }
      responses:
        "204":
          description: User and all dependent state deleted.
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403":
          description: Administrator privileges required, or the admin tried to delete their own account.
          x-error-codes: [forbidden, cannot_delete_self]
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
        "404":
          description: User not found.
          x-error-codes: [not_found]
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
        "409":
          description: The target is the last administrator account.
          x-error-codes: [last_admin]
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
        "429": { $ref: "#/components/responses/RateLimited" }
        "500": { $ref: "#/components/responses/InternalError" }
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Access token from POST /api/v1/login or POST /api/v1/refresh.
  parameters:
    ZoneName:
      name: name
      in: path
      required: true
      description: Zone name.
      schema: { type: string, maxLength: 64, example: homelab }
    NodeID:
      name: id
      in: path
      required: true
      description: ID of one of the caller's devices.
      schema: { type: integer, format: int64 }
  responses:
    Unauthorized:
      description: Missing, malformed or expired bearer token.
      x-error-codes: [unauthorized]
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorResponse" }
    ForbiddenAdmin:
      description: The authenticated account is not an administrator.
      x-error-codes: [forbidden]
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorResponse" }
    RateLimited:
      description: Global rate limit exceeded; retry shortly.
      x-error-codes: [rate_limited]
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorResponse" }
    InternalError:
      description: Unexpected server-side failure.
      x-error-codes: [internal_error]
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorResponse" }
  schemas:
    ErrorResponse:
      type: object
      description: Single error envelope used by every endpoint.
      required: [error, message]
      properties:
        error:
          type: string
          description: Machine-readable error code.
          example: validation_error
        message:
          type: string
          description: Human-readable sentence, safe to display.
          example: Invalid request body.
...
    LoginRequest:
      type: object
      required: [username, password]
      properties:
        username: { type: string, maxLength: 64, example: alice }
        password: { type: string, format: password, example: "CorrectHorse9" }
    LoginResponse:
      type: object
      required: [token, refresh_token]
      properties:
        token:
          type: string
          description: Short-lived JWT access token (send as `Authorization Bearer`).
          example: "xxxxx.yyyyy.zzzzz"
        refresh_token:
          type: string
          description: Long-lived opaque token for POST /api/v1/refresh; store securely.
          example: "c29tZS1vcGFxdWUtZXhhbXBsZS10b2tlbg"
    RefreshRequest:
      type: object
      required: [refresh_token]
      properties:
        refresh_token: { type: string, example: "c29tZS1vcGFxdWUtZXhhbXBsZS10b2tlbg" }
    RefreshResponse:
      type: object
      required: [token]
      properties:
        token: { type: string, description: Fresh JWT access token., example: "xxxxx.yyyyy.zzzzz" }
    LogoutRequest:
      type: object
      required: [refresh_token]
      properties:
        refresh_token: { type: string, example: "c29tZS1vcGFxdWUtZXhhbXBsZS10b2tlbg" }
    RegisterRequest:
      type: object
      required: [invite_code, username, password]
      properties:
        invite_code: { type: string, example: "EXAMPLE-INVITE-CODE" }
        username: { type: string, maxLength: 64, example: alice }
        password:
          type: string
          format: password
          description: 8-64 ASCII characters with at least one uppercase letter, one lowercase letter and one digit; no spaces.
          example: "CorrectHorse9"
    RegisterResponse:
      type: object
      required: [username, is_admin]
      properties:
        username: { type: string, example: alice }
        is_admin: { type: boolean, example: false }
    MeResponse:
      type: object
      required: [user_id, username, is_admin]
      properties:
        user_id: { type: integer, format: int64, example: 7 }
        username: { type: string, example: alice }
        is_admin: { type: boolean, example: false }
    ServerInfoResponse:
      type: object
      required: [public_key, endpoint, network, mtu]
      properties:
        public_key:
          type: string
          description: Server WireGuard public key (base64).
          example: "AAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAA="
        endpoint:
          type: string
          description: Public host:port clients dial over UDP.
          example: "vpn.example.com:51820"
        network:
          type: string
          description: VPN address pool in CIDR form.
          example: "100.127.0.0/16"
        mtu: { type: integer, example: 1420 }
    RegisterNodeRequest:
      type: object
      required: [name, wg_pubkey]
      properties:
        name: { type: string, maxLength: 64, example: laptop }
        wg_pubkey:
          type: string
          description: Device WireGuard public key (base64).
          example: "AAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAA="
    NodeResponse:
      type: object
      required: [id, name, ip, online]
      properties:
        id: { type: integer, format: int64, example: 3 }
        name: { type: string, example: laptop }
        ip: { type: string, description: Assigned VPN address., example: "100.127.0.7" }
        created_at: { type: string, description: RFC3339 creation time (omitted in some responses)., example: "2026-06-10T12:00:00Z" }
        online:
          type: boolean
          description: True when the last WireGuard handshake is younger than 3 minutes.
          example: true
        last_handshake: { type: string, description: RFC3339 time of the last handshake; omitted when never connected., example: "2026-06-10T12:34:56Z" }
    NodeListResponse:
      type: object
      required: [nodes]
      properties:
        nodes:
          type: array
          items: { $ref: "#/components/schemas/NodeResponse" }
    CreateZoneRequest:
      type: object
      required: [name, password]
      properties:
        name: { type: string, maxLength: 64, example: homelab }
        password:
          type: string
          format: password
          description: Zone join password, at least 8 characters.
          example: "zone-pass-example"
        node_id:
          type: integer
          format: int64
          description: Optional. One of the caller's device IDs to auto-join in the same operation; 0/omitted creates the zone without members.
          example: 3
    ZoneResponse:
      type: object
      required: [id, name, is_owner]
      properties:
        id: { type: integer, format: int64, example: 5 }
        name: { type: string, example: homelab }
        is_owner: { type: boolean, example: true }
    ZoneListResponse:
      type: object
      required: [zones]
      properties:
        zones:
          type: array
          items: { $ref: "#/components/schemas/ZoneResponse" }
    JoinZoneRequest:
      type: object
      required: [node_id, password]
      properties:
        node_id: { type: integer, format: int64, example: 3 }
        password: { type: string, format: password, example: "zone-pass-example" }
    LeaveZoneRequest:
      type: object
      required: [node_id]
      properties:
        node_id: { type: integer, format: int64, example: 3 }
    ChangeZonePasswordRequest:
      type: object
      required: [password]
      properties:
        password:
          type: string
          format: password
          description: New zone join password, at least 8 characters. Existing members stay.
          example: "new-zone-pass-1"
    ZoneMemberResponse:
      type: object
      required: [node_id, node_name, ip, owner]
      properties:
        node_id: { type: integer, format: int64, example: 3 }
        node_name: { type: string, example: laptop }
        ip: { type: string, example: "100.127.0.7" }
        owner: { type: string, description: Username of the device owner., example: alice }
    ZoneMembersResponse:
      type: object
      required: [members]
      properties:
        members:
          type: array
          items: { $ref: "#/components/schemas/ZoneMemberResponse" }
    CreateInviteResponse:
      type: object
      required: [code]
      properties:
        code: { type: string, example: "EXAMPLE-INVITE-CODE" }
        expires_at:
          type: string
          description: RFC3339 moment the code stops being redeemable; omitted when it never expires.
          example: "2026-06-11T12:00:00Z"
    InviteListItem:
      type: object
      required: [code, status, created_by, created_at, used_by, used_at]
      properties:
        code: { type: string, example: "EXAMPLE-INVITE-CODE" }
        status:
          type: string
          enum: [unused, used, expired]
          example: unused
        created_by: { type: string, nullable: true, example: admin }
        created_at: { type: string, example: "2026-06-10T12:00:00Z" }
        used_by: { type: string, nullable: true, example: null }
        used_at: { type: string, nullable: true, example: null }
        expires_at:
          type: string
          description: RFC3339 expiry; omitted when the code never expires.
          example: "2026-06-11T12:00:00Z"
    InviteListResponse:
      type: object
      required: [invites]
      properties:
        invites:
          type: array
          items: { $ref: "#/components/schemas/InviteListItem" }